Test_framework/requirements.txt
xlrd==1.1.0
pyyaml==3.12
aiohttp==3.3.2